requirements.txt
accelerate==0.28.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.12.3
bitsandbytes==0.49.0
build==1.3.0
cachetools==6.2.4
certifi==2022.12.7
charset-normalizer==2.1.1
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
datasets==2.18.0
dill==0.3.8
distro==1.9.0
docstring-parser==0.17.0
durationpy==0.10
exceptiongroup==1.3.1
fastapi==0.127.0
ffmpy==1.0.0
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2024.2.0
google-auth==2.45.0
googleapis-common-protos==1.72.0
gradio==4.19.2
gradio-client==0.10.1
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.21.0
humanfriendly==10.0
idna==3.4
importlib-metadata==8.7.1
importlib-resources==6.5.2
jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==34.1.0
lxml==6.0.2
markdown==3.5.2
markdown-it-py==4.0.0
markupsafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
mmh3==5.2.0
monotonic==1.6
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
narwhals==2.14.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.23.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandas==2.3.3
peft==0.10.0
pillow==10.4.0
posthog==3.3.1
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.0
pulsar-client==3.8.0
pyarrow==22.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.5
pydantic-core==2.41.5
pydub==0.25.1
pygments==2.19.2
pymupdf==1.24.0
pymupdfb==1.24.0
pyparsing==3.3.1
pypika==0.48.9
pyproject-hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
pyyaml==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
ruff==0.14.10
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==2.5.1
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.1
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.15.2
tomli==2.3.0
tomlkit==0.12.0
tqdm==4.67.1
transformers==4.38.0
trl==0.8.1
typeguard==4.4.4
typer==0.20.1
typer-slim==0.20.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tyro==1.0.3
tzdata==2025.3
urllib3>=2.0.0,<2.4.0
uvicorn==0.40.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==11.0.3
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
